requirements.txt
plexapi
pylast
Flask
python-dotenv
pyyaml